Make all copyright tests data driven
 * a few dnaling test files need to be added to the tests
   They are in the "dnagling" directory

+++ tests/cluecode/cluecode_assert_utils.py
@@ -30,18 +30,17 @@ import codecs
 from collections import OrderedDict
 import os
 from os.path import abspath
+from os.path import exists
 from os.path import join
+from unittest.case import expectedFailure
 
 import attr
 
 import cluecode.copyrights
+from commoncode import saneyaml
 from commoncode.testcase import FileDrivenTesting
 from commoncode.text import python_safe_name
-from commoncode import fileutils
-from commoncode import saneyaml
-from commoncode.fileutils import file_extension
-from commoncode.fileutils import parent_directory
-from commoncode.fileutils import file_name
+
 
 """
 Data-driven Copyright test utilities.
@@ -51,70 +50,15 @@ test_env = FileDrivenTesting()
 test_env.test_data_dir = os.path.join(os.path.dirname(__file__), 'data')
 
 
-def expectedFailure(func):
-    from unittest.case import expectedFailure
-    wrapper = expectedFailure(func)
-    wrapper.__wrapped__ = func
-    wrapper.is_expected_failure = True
-    return wrapper
-
-
-def check_detection(expected, test_file,
-                    what='copyrights',
-                    notes=None,
-                    expected_failure=False):
-    """
-    Run detection of copyright on the `test_file`, checking the
-    results match the expected list of values.
-
-    `test_file` is either a path string or an iterable of text lines.
-    """
-    test_file = test_env.get_test_loc(test_file)
-    parent = parent_directory(test_file)
-    nm = file_name(test_file)
-    data_file = join(parent, nm + '.yml')
-    tst = CopyrightTest()
-    tst.test_file = test_file
-    tst.what = what
-    tst.data_file = data_file
-    tst.notes = notes
-    tst.expected_failure = expected_failure
-    tst.dump()
-
-
-def check_detection_orig(expected, test_file,
-                    what='copyrights',
-                    notes=None,
-                    expected_failure=False):
-    """
-    Run detection of copyright on the `test_file`, checking the
-    results match the expected list of values.
-
-    `test_file` is either a path string or an iterable of text lines.
-    """
-    test_file = test_env.get_test_loc(test_file)
-
-    copyrights, authors, years, holders = cluecode.copyrights.detect(test_file)
-    results = {
-        'copyrights': copyrights,
-        'authors': authors,
-        'years': years,
-        'holders': holders
-    }
-
-    result = results[what]
-    assert expected == result
-
-
 @attr.s(slots=True)
 class CopyrightTest(object):
     """
     A copyright detection test is used to verify that copyright detection works
     correctly
 
-    It consists of two files with the same base name: a .yml file with test data
-    and a test file with any other extension that needs to be tested for
-    detection.
+    It consists of two files with the same file name: a .yml file with test data
+    and a test file with any other extension (and the same name whenremoving the
+    .yml extension) that needs to be tested for detection.
 
     The following data are loaded based on or from the .yml file:
      - a test file to scan for copyrights (based on file name convenstions),
@@ -128,7 +72,7 @@ class CopyrightTest(object):
     """
     data_file = attr.ib(default=None)
     test_file = attr.ib(default=None)
-    what = attr.ib(default=None)
+    what = attr.ib(default='copyrights')
     expectations = attr.ib(default=attr.Factory(list))
     expected_failure = attr.ib(default=False)
     notes = attr.ib(default=None)
@@ -144,17 +88,19 @@ class CopyrightTest(object):
         """
         Serialize self to an ordered mapping.
         """
-        filtered = [(name, field) for name, field in attr.fields(CopyrightTest)
-                    if b'_file' in name]
+        filtered = [field for field in attr.fields(CopyrightTest)
+                    if b'_file' in field.name]
         fields_filter = attr.filters.exclude(*filtered)
         data = attr.asdict(self, filter=fields_filter, dict_factory=OrderedDict)
         return OrderedDict([(key, value) for key, value in data.items() if value])
 
-    def dump(self):
+    def dump(self, check_exists=False):
         """
         Dump a representation of self to a .yml data_file in YAML block format.
         """
         as_yaml = saneyaml.dump(self.to_dict())
+        if check_exists and os.path.exists(self.data_file):
+            raise Exception(self.data_file)
         with codecs.open(self.data_file, 'wb', encoding='utf-8') as df:
             df.write(as_yaml)
 
@@ -163,25 +109,49 @@ def load_copyright_tests(test_dir=test_env.test_data_dir):
     """
     Yield an iterable of CopyrightTest loaded from test data files in test_dir.
     """
-    # first collect files with .yml extension and files with other extensions
-    # in two maps keyed by file base_name
+    # first collect files with .yml extension and files with other no extensions
+    # extension in two maps keyed by the test file path
     data_files = {}
     test_files = {}
-    for top, _, files in os.walk(test_dir):
-        for yfile in files:
-            if yfile. endswith('~'):
-                continue
-            base_name = fileutils.file_base_name(yfile)
-            file_path = abspath(join(top, yfile))
-            if yfile.endswith('.yml'):
-                assert base_name not in data_files
-                data_files[base_name] = file_path
-            else:
-                assert base_name not in test_files
-                test_files[base_name] = file_path
+    dangling = set()
+
+    def collect(subdir):
+        for top, _, files in os.walk(join(test_dir, subdir)):
+            for yfile in files:
+                if yfile. endswith('~'):
+                    continue
+                file_path = abspath(join(top, yfile))
+
+                if yfile.endswith('.yml'):
+                    data_file_path = file_path
+                    test_file_path = file_path.replace('.yml', '')
+                else:
+                    test_file_path = file_path
+                    data_file_path = test_file_path + '.yml'
+
+                if not exists(test_file_path):
+                    dangling.add(data_file_path)
+                if not exists(data_file_path):
+                    dangling.add(test_file_path)
+                data_files[test_file_path] = data_file_path
+                test_files[test_file_path] = test_file_path
+
+    for d in ('copyrights', 'ics', 'holders', 'authors', 'years'):
+        collect(d)
+
+    if dangling:
+        print('Dangling missingcopyright test files')
+        for o in sorted(dangling):
+            print(o)
+            raise Exception('Dangling missingcopyright test files')
 
     # ensure that each data file has a corresponding test file
     diff = set(data_files.keys()).symmetric_difference(set(test_files.keys()))
+    if diff:
+        print('orphaned copyright test files')
+        for o in sorted(diff):
+            print(o)
+
     assert not diff, ('Orphaned copyright test file(s) found: '
                       'test file without its YAML test descriptor '
                       'or YAML test descriptor without its test file.')
@@ -193,20 +163,14 @@ def load_copyright_tests(test_dir=test_env.test_data_dir):
         yield CopyrightTest(data_file, test_file)
 
 
-def make_copyright_test_function(
-        expected, test_file, test_data_file, test_name,
-        what='copyrights', expected_failure=False):
+def make_copyright_test_function(test, test_name):
     """
     Build and return a test function closing on tests arguments.
     """
-    if isinstance(test_name, unicode):
-        test_name = test_name.encode('utf-8')
-
-    assert isinstance(expected, list)
 
     def closure_test_function(*args, **kwargs):
         try:
-            copyrights, authors, years, holders = cluecode.copyrights.detect(test_file)
+            copyrights, authors, years, holders = cluecode.copyrights.detect(test.test_file)
             results = OrderedDict([
                 ('copyrights', copyrights),
                 ('authors', authors),
@@ -214,40 +178,36 @@ def make_copyright_test_function(
                 ('holders', holders),
             ])
 
-            result = results[what]
-            assert expected == result
+            result = results[test.what]
+            assert test.expectations == result
         except:
             # On failure, we compare against more result data to get additional
             # failure details, including the test_file and full results
             # this assert will always fail and provide a more detailed failure trace
-            assert expected == results + [test_name, 'test file: file://' + test_file]
+            assert test.expectations == results + [test_name, 'test file: file://' + test.test_file]
 
     closure_test_function.__name__ = test_name
     closure_test_function.funcname = test_name
 
-    if expected_failure:
+    if test.expected_failure:
         closure_test_function = expectedFailure(closure_test_function)
 
     return closure_test_function
 
 
-def build_tests(copyright_tests, clazz):
+def build_tests(copyright_tests, clazz, test_data_dir=test_env.test_data_dir):
     """
     Dynamically build test methods from a sequence of CopyrightTest and attach
     these method to the clazz test class.
     """
     for test in copyright_tests:
-        tfn = test.test_file_name
-        test_name = 'test_copyright_%(tfn)s' % locals()
+        tfn = test.test_file
+        tfn = tfn.replace(test_data_dir, '').strip('\/\\')
+        test_name = 'test_%(tfn)s' % locals()
         test_name = python_safe_name(test_name)
-
+        if isinstance(test_name, unicode):
+            test_name = test_name.encode('utf-8')
         # closure on the test params
-        test_method = make_copyright_test_function(
-            test.expectations, test.test_file, test.data_file,
-            test_name=test_name,
-            what=test.what,
-            expected_failure=test.expected_failure,
-        )
-
+        test_method = make_copyright_test_function(test, test_name)
         # attach that method to our test class
         setattr(clazz, test_name, test_method)
